platform cleanup px4_log and dependency mess
 - cmake add px4_modules interface library for linking all modules
