+++ boards/intel/aerofc-v1/src/CMakeLists.txt
@@ -32,7 +32,7 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	init.c
 	led.c
 	spi.c
