+++ boards/nxp/fmuk66-v3/src/CMakeLists.txt
@@ -31,7 +31,7 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	autoleds.c
 	automount.c
 	can.c
