+++ boards/omnibus/f4sd/src/CMakeLists.txt
@@ -31,7 +31,7 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	init.c
 	led.c
 	spi.c
