+++ boards/px4/cannode-v1/src/CMakeLists.txt
@@ -31,7 +31,7 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	buttons.c
 	can.c
 	init.c
