+++ boards/px4/fmu-v3/src/manifest.c
@@ -47,10 +47,12 @@
  * Included Files
  ****************************************************************************/
 
-#include <px4_config.h>
+#include <nuttx/config.h>
+#include <board_config.h>
+
 #include <stdbool.h>
+
 #include "systemlib/px4_macros.h"
-#include "px4_log.h"
 
 /****************************************************************************
  * Pre-Processor Definitions
@@ -134,7 +136,7 @@ __EXPORT px4_hw_mft_item board_query_manifest(px4_hw_mft_item_id_t id)
 		}
 
 		if (boards_manifest == px4_hw_mft_list_uninitialized) {
-			PX4_ERR("Board %4x is not supported!", ver_rev);
+			syslog(LOG_ERR, "[boot] Board %4x is not supported!\n", ver_rev);
 		}
 	}
