+++ boards/px4/fmu-v4/src/CMakeLists.txt
@@ -31,7 +31,7 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	can.c
 	init.c
 	led.c
