+++ boards/px4/io-v2/src/CMakeLists.txt
@@ -31,7 +31,8 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	init.c
 	timer_config.c
 )
+add_dependencies(drivers_board prebuild_targets)
