+++ boards/px4/raspberrypi/src/CMakeLists.txt
@@ -31,6 +31,6 @@
 #
 ############################################################################
 
-px4_add_library(drivers_board
+add_library(drivers_board
 	${PX4_SOURCE_DIR}/src/drivers/boards/common/board_identity.c # TODO: this is horrible and should be fixed
 )
