+++ msg/CMakeLists.txt
@@ -210,5 +210,5 @@ add_custom_command(OUTPUT ${uorb_sources}
 	VERBATIM
 	)
 
-add_library(uorb_msgs ${uorb_sources})
+add_library(uorb_msgs EXCLUDE_FROM_ALL ${uorb_sources})
 add_dependencies(uorb_msgs prebuild_targets uorb_headers)
