+++ platforms/nuttx/src/px4_layer/CMakeLists.txt
@@ -32,20 +32,17 @@
 ############################################################################
 
 # skip for px4_layer support on an IO board
-if (NOT ${PX4_BOARD} MATCHES "px4_io")
-
-	add_library(px4_layer
-		px4_nuttx_tasks.c
-		px4_nuttx_impl.cpp
-		px4_init.cpp
+add_library(px4_layer
+	px4_nuttx_tasks.c
+	px4_nuttx_impl.cpp
+	px4_init.cpp
+)
+target_link_libraries(px4_layer
+	PRIVATE
+		nuttx_apps # up_cxxinitialize
+		nuttx_sched
+		drivers_boards_common_arch
+		px4_platform_log
+		px4_platform_uorb
+		prebuild_targets
 	)
-	target_link_libraries(px4_layer
-		PRIVATE
-			nuttx_apps # up_cxxinitialize
-			nuttx_sched
-			drivers_boards_common_arch
-		)
-else()
-	add_library(px4_layer ${PX4_SOURCE_DIR}/src/platform/empty.c)
-endif()
-add_dependencies(px4_layer prebuild_targets)
