+++ platforms/qurt/CMakeLists.txt
@@ -35,7 +35,7 @@ if ("${QURT_ENABLE_STUBS}" STREQUAL "1")
 		${PX4_BINARY_DIR}/platforms/qurt/px4muorb_skel.c
 	)
 
-	target_link_libraries(px4 PRIVATE ${module_libraries} ${df_driver_libs})
+	target_link_libraries(px4 PRIVATE px4_modules ${df_driver_libs})
 
 else()
 	# Generate the DSP lib and stubs but not the apps side executable
@@ -46,7 +46,7 @@ else()
 			${PX4_BINARY_DIR}/apps.cpp
 		LINK_LIBS
 			modules__muorb__adsp
-			${module_libraries}
+			px4_modules
 			${df_driver_libs}
 			m
 		)
