+++ src/drivers/boards/common/CMakeLists.txt
@@ -32,29 +32,21 @@
 ############################################################################
 
 # common board drivers (currently only for nuttx fmu boards)
-if ((${PX4_PLATFORM} MATCHES "nuttx") AND NOT ${PX4_BOARD} MATCHES "io")
+add_library(drivers_boards_common
+	board_crashdump.c
+	board_dma_alloc.c
+	board_gpio_init.c
+	)
 
-	add_library(drivers_boards_common
-		board_crashdump.c
-		board_dma_alloc.c
-		board_gpio_init.c
-		)
-
-	if (${CONFIG_ARCH_CHIP} MATCHES "kinetis")
-		add_subdirectory(kinetis)
-	elseif (${CONFIG_ARCH_CHIP} MATCHES "stm32")
-		add_subdirectory(stm32)
-	endif()
-
-	target_link_libraries(drivers_boards_common
-		PRIVATE
-			drivers_boards_common_arch
-			nuttx_arch # bbsram
-			nuttx_mm # dma gran alloc
-			systemlib
-		)
-
-else()
-	add_library(drivers_boards_common INTERFACE)
-	add_library(drivers_boards_common_arch INTERFACE)
+if (${CONFIG_ARCH_CHIP} MATCHES "kinetis")
+	add_subdirectory(kinetis)
+elseif (${CONFIG_ARCH_CHIP} MATCHES "stm32")
+	add_subdirectory(stm32)
 endif()
+
+target_link_libraries(drivers_boards_common
+	PRIVATE
+		drivers_boards_common_arch
+		nuttx_arch # bbsram
+		nuttx_mm # dma gran alloc
+	)
