+++ src/modules/muorb/adsp/px4muorb.cpp
@@ -33,15 +33,14 @@
 
 #include "px4muorb.hpp"
 #include "uORBFastRpcChannel.hpp"
-#include "uORBManager.hpp"
+#include <uORB/uORBManager.hpp>
 
 #include <px4_middleware.h>
 #include <px4_tasks.h>
 #include <px4_posix.h>
 #include <dspal_platform.h>
-#include "px4_log.h"
-#include "uORB/topics/sensor_combined.h"
-#include "uORB.h"
+
+#include <uORB/uORB.h>
 #include <parameters/param.h>
 #include <shmem.h>
