+++ src/platform/CMakeLists.txt
@@ -31,7 +31,8 @@
 #
 ############################################################################
 
-add_subdirectory(common)
+add_subdirectory(common EXCLUDE_FROM_ALL)
 
 add_subdirectory(cdev EXCLUDE_FROM_ALL)
+add_subdirectory(log EXCLUDE_FROM_ALL)
 add_subdirectory(uORB EXCLUDE_FROM_ALL)
