+++ src/platform/common/CMakeLists.txt
@@ -31,20 +31,11 @@
 #
 ############################################################################
 
-set(SRCS)
-
-if (NOT "${PX4_PLATFORM}" MATCHES "qurt" AND NOT "${PX4_BOARD}" MATCHES "io-v2")
-	list(APPEND SRCS
-		px4_log.c
-		)
-endif()
-
 add_library(px4_platform
 	module.cpp
 	px4_getopt.c
 	px4_cli.cpp
 	shutdown.cpp
-	${SRCS}
 	)
 add_dependencies(px4_platform prebuild_targets)
 add_subdirectory(work_queue)
