+++ src/platform/log/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2018 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -30,11 +30,12 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 ############################################################################
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wno-error")
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-error")
 
-add_library(drivers_board
-	init.c
-	${PX4_SOURCE_DIR}/src/drivers/boards/common/board_identity.c # TODO: this is horrible and should be fixed
+add_library(px4_platform_log EXCLUDE_FROM_ALL
+	px4_log.c
 )
+add_dependencies(px4_platform_log prebuild_targets)
 
+if("${PX4_PLATFORM}" MATCHES "posix")
+	target_link_libraries(px4_platform_log PRIVATE px4_daemon)
+endif()
