+++ src/platform/log/px4_log.c
